Move entity list tests to service overview tests
Prior to this change, the entity overview was tested in the entity overview detail page.

This change moves those test to the service overview page as the entity overview detail page is being deleted.  It also adds some new handlers to reflect the new structure of the page & updates translations where necessary.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607