Style graph section & navigation
Prior to this change, the graph section had no title & could not be minimized.  Also the navigation for a service was a navbar.

This change adds a title to the graph section & makes it minimizable for those who find it takes up too much screenspace.  Additionally, the change moves the navigation for the service in two parts:
- edit & delete action move to the title of the service (only for admins)
- the privacy link moves to the service graph title

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607