Add back to top links to service-cards
Prior to this change, a user had no way of scrolling back to top after each card.  Depending on the number of services and the number of entities they have, that can mean a lot of scrolling to get back to top.

This change adds an affordance for users to scroll back to the top at the end of each service-card

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607