Add a privacy notification
Prior to this change, the graph was supposed to nudge people to fill in the privacy questions.  This was no longer viable as the graph was moved below the table & it also didn't work.

This change adds 2 notifications alerting the user that the privacy questions need to be filled out.  Both are styled to be noticed immediately.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165965607