Ensure copy form id is editable
Prior to this change, when copying a oicdng or oicdng-rs entity, the id was not editable.

This change adds a check to see whether or not the edit form is due to a copy action.  If it is a copy action, the id field is editable.  Else it is not.

See the story in pivotal for more info: https://www.pivotaltracker.com/story/show/165599582