Prevent RS overwrites when outside of team
When service desk employees connect RS to an RP, they can actually
connect a RS of a team outside of the RP. SPD would not consider this a
valid option and yanked them out of the metadata pushed back to Manage.
This change does allow for this behavior. By comparing the team names
from the entities, and keeping the ones from outside the enities own
organization.