Show attribute without motivation as requested
When in manage no motivation is specified for a SPD tracked attribute,
previously we would not show it on the form. It is shown now, but
without motivation. This prevents us from saving it as if it didn't
exist (effectivly removing it from manage). But this does force the user
to fill the motivation in order to publish the form.