Show new production entity link for admins
Prior to this change, admins could not see the "new production entity" link.  Added to that when the label text was semi-long the hover effect was a tad ugly.

This change ensures that admins can see the link & that the hover effect is ok for a longer label.

Found when testing release 2.8