Alter form validation to check error length
Prior to this change, the validation would always throw an error for the create entity modal when submitting a test entity.  This is because if the form is not "submitted" it is considered invalid.

This change alters the validation to check whether there are actual form errors.  If none are found the form is considered valid.

Following the boyscout rule the ids for the titles were also made unique to the modal so they pass html validation.

Found when testing release 2.8