Ensure privacy link is visible for normal users
Prior to this change, the privacy link was only visible to admin users.  Meaning that once the questions were answered, you could no longer go to the privacy questions.

This change ensures the link is now also visible for normal users (as was the original intent).  Additionally it also adds a link to the privacy questions in the tooltip for the status summary.

Found when testing release 3.0