Repair faulty status implementation
Previously the status was added to the Jira Issue solution. This status
is used to figure out the entity status.

Problem is that the status id was not yanked from the Jira status field,
the entire status object was set on the type restricted Issue
constructor.

While at it, a similar development issue was fixed.

https://www.pivotaltracker.com/story/show/177659652