Update the Organization names when loading entity
The service organization name is leading in the determination of the org
name en/nl. The entity might be created before the org name was set on
the service. By updating the Organization (VO) names upon loading the
enity, we ensure every enitty is set with the correct name. This also
ensures every entity we tocuh afterwards wil have the correct names in
Manage.

https://www.pivotaltracker.com/story/show/177701148