Hide edit for PE on removal requested
Prior to this change, the edit option was visible for entities that had the status of removal requested.  Upon clicking the edit button, an access denied screen was shown.  This behaviour was right, but the option to click edit should not be present.

This change removes the edit option.

Pivotal ticket: https://www.pivotaltracker.com/story/show/178034009