Add a global site notice
Prior to this change, there was no way to alert the user with temporary content.  For example: what changed in the last release.

This change adds a global site notice.  It uses two parameters to set the allowed html & whether or not to display the notice.  The content of the notice itself can be set in the translation.

Pivotal ticket at: https://www.pivotaltracker.com/story/show/176902060