Lowercase client ids and redirect url
Prior to this change, Client Ids and redirect url could be a mix of lowercase and uppercase.

This change lowercases the redirect urls and the client ids in the saveOidcNgEntity.  It also applies the boyscout rule and fixes an error when viewing the oidc entities.

Pivotal ticket @: https://www.pivotaltracker.com/story/show/173276689