Validate entity id for SAML
Prior to this change, the entity id was not validated for uniqueness with SAML entities.  This lead to an indistinct error when publishing.

This change applies the correct validation for uniqueness.  This ensures the error is shown on the create page with a more helpful text.

Pivotal ticket @: https://www.pivotaltracker.com/story/show/172022288