Check for the full urn on the edit command
On the edit command it was still possible to enter a teamname which wasn't in urn format.  This lead to problems when authorising at login.

This change ensures that we check whether the urn format is the correct form and if not throw an error.

See pivotal @https://www.pivotaltracker.com/story/show/180260470/comments/227984970