Retrieve only Unresolved issues from Jira
During a peer programming session with Bart, we encountered that the
request delete of prod entity status was not reverted back when the
issue status in Jira was set to 'resolved'. Resulting in the ticket
retaining its faulty status. This feature was encountered when users
unintentionally clicked the 'remove prod entity' button. And the support
team verified the action with the user, resulting in the resolution of
the ticket (without any action). This should have resulted in the
entityt status reverting back to 'published'.

By simply filtering the jira results to only show the Unresolved items
we resolved this issue. With the additional benefit of only retrieving
the active tickets at that point. Decreasing the processing time and
resource hogging.

See: https://www.pivotaltracker.com/story/show/176244612