Implement Comparable interface on ManageEntity
And all its parts it is made off. This interface puts a asArray method
on all the entity parts. The manageEntity itself triggers the calling of
that method via its diff method.

The diff method takes another ManageEntity that it will be compared to.
The differences are returned as a Manage-esque json serializable array

See: https://www.pivotaltracker.com/story/show/182413030