Add isSet method to Contact value object
Instead of introducing a new 'NullContact' I opted for adding an isSet
method to the Contact VO. Why? Contacts would always  be created for
each contact type. This made comparing the contacts difficult and would
result in removing actual contacts from manage. As a contact with null
values are interpreted by Manage as them having to be removed.

By using the isSet method on the diff & json generator this allows us to
easily skip them in the publication process.