Create a EntityChangeRequestCommand and send it
This command is created when editing a published prod entity.

It is handled in a newly created handler (copy paste of the entity
publish entity). The handler is not yet pruned and rewritten to fit the
new use case yet. That is up next. Most of the dependencies have been
left in place. As lateron we are actually going to use Jira, the mailer
and the other services to notify the service desk.