Generate ARP when actual changes where applied
Previously we would NOT diff the AttributeList. We now do that, not to
get the data we include on the Manage API payload. But to get an
indication on whether or not we need to actually add the ARP on the
payload. If we have a change, we just provide the entire arp. If no
changes where made, we refrain from adding the ARP on the payload.