Remove hard coded attribute references from view
Thet detail template no longer lists all attributes seperately. They are
now just loaded from the EntityDetail view object. By accessing its
attributes property. The next commit will show how that collection comes
to be.

The EntityDetailAttribute is a small peak into that solution. It shows
the view object that represents the attribute. Its an aggegation of the
Manage entity value for the attribute (the motivation) and the
information popup text (found in the attributes.json) and a label,
depicting the name of the attribute.