Allow searching for a single attribute
Over to the repository side of the application. This infrastructural
change was required to allow us to read attribute details for display on
the entity detail overview. The approach here is that a manage entity
has 0..n attributes in its arp. Those attributes need to be translated
to the attributes that make sense to us in SP Dashboard.

The Attribute service (next commit) calls the repository for each
attribute that it finds on a Manage entity.

If that entity is known to us (it is in the attributes.json). Then we
return that attribute Dto. If not, we return a NullAttribute. Both
implement the AttributeInterface.

The attributes.json was expanded with entity detail specific
configuration. Each of those attributes can also have a label and an
information popup value. They are handled and parsed the same as the `form`
attribute information.