Produce entity detail attrs in AttributeService
As discussed in the previous commit. The Attribute Service utilizes the
attribute repo to find specific details of a given attribute.

The service method combines the data gathered from the repository with
that of the corresponding Manage Entities attribute list. Combining the
view data required. Which is the motivation (from manage enitty) and the
UI elements (label and information popup) which is yanked from the
attributes.json.