Integrate the attribute detail changes
The four previous commits showed how individual parts of the application
where prepared for the dynamic display of an entities attributes.

This commit integrates those changes. The main concept to do so is by
introducing a factory that is tasked with creating the EntityDetail.
Previously, the EntityDetail view object would construct itself with a
static constructor. Now that we need to enrich its attribute data,
having a factory with those added dependencies makes more sense. An
added benefit was that injection of the playground uri's was no longer
required on the controller.

The factory calls the attribute services `createEntityDetailAttributes`
method. Effectively merging the Manage data with that collected in our
attributes.json data source.