Read the exclude on from property and interpret it
By reading the property from the dto, the Attribute value object can now
also be made aware of that setting. The value object received a
isExcluded method. Checking if a given protocol should be excluded from
view on a form.

This was integrated into the Saml and Oidc entity forms.