Transform Jira reporter name to custom field
The Reporter in Jira needs to be a known Jira user for the organisation
that subscribes to Jira. In this case we fill it with a name/mail
address of a Service contact person. And that person wil most probably
not be known in Jira. Yielding a big fat error.

The solution is to use a custom field (that is already provisioned in
the SURF Jira setup). That field is then filled with the Name/Email
combination.

This goal was reached by following the existing pattern that was already
in place for the ManageId and the EntityId custom fields.

A fixup for: https://github.com/SURFnet/sp-dashboard/pull/523
See https://www.pivotaltracker.com/story/show/179749372/comments/233715744