Remove flash message from prod edit action
When a prod entity is saved (change request) sucesfully, we used to add
a flash message. Showing the user the save action was a succes. This is
no longer the case, as this action would result in showing a thank you
page. And having both a flash message and a thank you page is a little
too much.

See: https://www.pivotaltracker.com/story/show/182246029/comments/233754893