Filter out resolved Jira tickets
When finding Jira tickets by manage id and an issue type (when looking
for a specific ticket for a specific entity). The `resolution` was not
taken into account. Resulting in opening no new ticket when a ticket was
previously resolved in Jira.

See: https://www.pivotaltracker.com/story/show/182411807/comments/233832245