Handle grants and redirect urls correctly
When diffing entity changes, the grants and redirect urls must be
included on the 'add all values' and not in a diff manner. I previously
implemented this. But it was added to the diff also when no actual
changes where applied to the data.

By only adding them when the diff result yielded a result. I worked
around that problem.

The test shows the different combinations. And the expected outcomes.
That should help in getting a better grasp of the correct behavior.