Improve readability and error handling publish entity
Different situations are covered when creating and
changing entities. These include creating an entity,
whether or not based on a copy, or editing an existing
entity. In addition, the environment is decisive:
test, production or otherwise
In case of a new (copy) entity, a connection request must
also be submitted. These situations and error handling
are now a bit more streamlined.