Simplyfy log setup
And update the php.ini to ensure we use stdout as the log location. To
do so, the docker container setup was changed slightly. The default
php.ini files, proposed by php.net are used for prod and dev php-fpm
containers a baseline config. Custom overrides of the monolog config
have been removed. And the monolog config is now configured to function
as expected in a docker environment. It logs to stderr