Upgrade Symfony webtests to utilize Panther (#576)
* Setup Panther for docker

* Setup Panther for docker

* Use test-browser as a headless chrome browser

* Docker compose: Remove extra hostname: Resolving works automatically

* Setup OSX session chrome fails

* Setup Panther client to run @ test env

* Use panther client for testing

* Downgrade version to run at t08 vm

* Initialize response

* Prepare for panther using client

* Append debug info at EOF

* Write and repair CreateConnectionRequestTest

Found one bug, where duplicate institutions where no longer
validatedd

* Repair CreatePrivacyQuestionsTest

In the process added a helper method to check checkboxes and a scrolling
feature in the DebugFile debug suite

* Repair EditPrivacyQuestionsTest

* Repair EditServiceTest

* Repair EntityAclTest

* Repaired EntityCopyTest

Removed a micky mouse test that had little or no added value.

* Repair EntityCreateClientCredentialsClientTest

1. Added sef::stofWebServer to the setup method of the tests, this
   increases stability of individual test runs
2. Removed some test related debug code

* Repair EntityCreateOidcngResourceServerTest

* Repair EntityCreateOidcngTest

* Repair EntityDeleteTest

While at it, the fake jira repo now also stores if it should simulate a
Jira outage. That state was lost between setting up the test and
actually runing the test in browser.

* Repair EntityDetailTest

* Repair EntityEditTest

* Repair IdentityHeaderTest

* Repair ServiceAdminOverviewTest

* Repair ServiceCreateTest

Removed the three last tests as they either did not add any added value,
or even the buisiness rules had changed and the test was modified to
behave different to what it advertised.

* Repair ServiceDeleteTest

* Repair ServiceOverviewTest

While at it: added support for multi service logins (where the assertion
contains multiple isMemberOf attribute values that are connected to one
of our services)

And also ensure logging output is sent to stdout. That way we can see
the app log in the docker logs

* Repair ServiceSwitcherTest

* Repair first 3 tests

They had not yet implemented the new login method

* Repair unit and integration tests

Some where broken because fixtures changed in rewriting the webtests

* Disable composer script timeout

* Silence dump-env compose script

* Migrate phpunit.xml config file to v9

* Docker: Copy migrations directory

* Docker: Touch a .env file, and add logging.yml to log to stderr

* Docker httpd: Copy the build dir to the right location

* Docker: Overwrite monolog.yaml with our logging config

* Update Teams Delete Service di config

* Make EntityCreateConnectionRequestCtrl SF5 proof

* Remove DebugFile references

* Optimize Dockerfile and ensure write on test db

* Repair ServiceDeleteTest

* Simplyfy log setup

And update the php.ini to ensure we use stdout as the log location. To
do so, the docker container setup was changed slightly. The default
php.ini files, proposed by php.net are used for prod and dev php-fpm
containers a baseline config. Custom overrides of the monolog config
have been removed. And the monolog config is now configured to function
as expected in a docker environment. It logs to stderr

* Import missing classes

The ManageEntity and the Constants where not properly imported.
Resulting in a 500 error

See: https://www.pivotaltracker.com/story/show/184712499

* Show client secret modal when creating prod oidc

* Update Symfony dependencies

* Disable deprecation warnings on unit tests

* Update Yarn dependencies

---------

Co-authored-by: Bart Geesink <bart.geesink@surf.nl>
Co-authored-by: Michiel Kodde <mkodde@ibuildings.nl>