Configure feature flag in yaml
The jira_enable_test_mode feature flag should be configured as a yaml
parameter as this can not be read during compile time from the .env
file. That would defeat the purpose of the .env files being overridable
during runtime.

That's why services.yml files where added for all environments. Allowing
to override the setting for any environment to enable/disable the jira
test mode.