Display custom authnFailed error page
In order to show the unknown service error page (when the team id is not
matching up to any of the services kwown to SPD). I needed to tie in to
the exception listener service. As SF would just render a generic 500
page when the ServiceNotFound exception remained uncatched. The listener
now checks if that exception is handled. If so, it renders the fancy
error page.

The authnFailed page was not yet displayed when authn fails for
whatever reason. You would expect the exception listener would be used
for this purpose, but the FailureHandler is used for this..