Prevent double actions
When clicking certain buttons in the teams screens, events would be
triggered multiple times. This was already fixed in #600. But other
construct also used the mousedown event. That was fixed in this commit
for the change member the resent invitation listeners. The issue was
described in the storie mentioned below.

See: https://www.pivotaltracker.com/story/show/185785974
See: https://github.com/SURFnet/sp-dashboard/pull/600/files