Move TypeOfServiceRepositoryFromConfig to domain
It having a coupling with filesystem does not have to mean it is not
part of the domain. That was a fault in my reasoning. Also to more
easily reuse the repository. I decoupled the repo file path from DI and set
it in the Contstants.

That way we can also use it in the Coin value object factory method