Sharpen the EntityDiff algo
The check would be type unspecific. That resulted in situations where "" would
be considered null. That resulted in the new coin value to never be set
for existing entities. As null would also evaluate to be true when the
new value for the value is `false`. As php thinks false == null when not testing
in a strict manner.