Hide nameIdFomrat when unspecified
When the 'subjectType' or 'nameIdFormat' is unspecified. The matching form field
is not displayed on the form. That way we do not allow the user to
overwrite this value in Manage (which is explicitly set for that SP by a
product owner)

To test this, the mock SP storage needed to track the nameIdFormat value
as a manage entity. That requires quite some test code to be edited..

See: https://www.pivotaltracker.com/story/show/185952423