Address some deprecations
Prior to this change, the ServiceOverviewTest would load the fixtures twice.
This would make doctrine complain about primary keys being forcefully reused on new enities.
This change does not load the fixtures twice, resolving the issue.

Prior to this change, doctrine would note that we need to make a decision whether to use savepoints or not.
This change explicitly enables savepoints, as Innodb engine and sqlite support savepoints.
This was always the default behavior, and by explicitly setting it to
true, we keep it that way. Thil will tell doctrine that it can use
nested database transactions.

Also rewrite test assertion to not rely on deprecated Hamcrest bundle.