PHPCS: tweak rules: todo comments
Prior to this change, todo comments would not trigger phpcs.
This change makes the pipeline red on 'todos'.
In order to get this to work, the phpcs runtime needed to be changed, because if warning were not 'severe' enough, nothing happened.
This change removes ignoring 'low prio' warnings, as nobody saw then, and the pipeline was green.
It was only used for soft line length enforcement, but if the pipeline does not turn red, it does nothing, so its removed.