Improve UX of Create connection request form
Prior to this change, when you enterd the form data, it was unclear you needed to press the + button in order to 'add' it.
This change shows the list headers + a clarifying text to clarify this.
The clarifying text should only be shown if there are no items in the list.

Resolves https://github.com/SURFnet/sp-dashboard/issues/1201