Fix: Don't reset contractual base on production entities
Prior to this change, when a change request was sent, the contractualBase would be set to null, because it is not present in the form.
This change updates the contractual base if necessary on update.

The expected behaviour is:
When the field is not present, add it with the correct value
When the field is present and the value is incorrect update it
When the field is present and the value is correct, leave it untouched

Resolves https://github.com/SURFnet/sp-dashboard/issues/1343